[ART] Refactor POA holder serializer
