+++ modules/accredited_representative_portal/app/serializers/accredited_representative_portal/power_of_attorney_request_serializer/individual_power_of_attorney_holder_serializer.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module AccreditedRepresentativePortal
+  class PowerOfAttorneyRequestSerializer
+    class IndividualPowerOfAttorneyHolderSerializer < ApplicationSerializer
+      attribute :type do |poa_holder|
+        "accredited_#{poa_holder.individual_type}"
+      end
+
+      attribute :full_name do |poa_holder|
+        parts = [
+          poa_holder.first_name,
+          poa_holder.middle_initial,
+          poa_holder.last_name
+        ]
+
+        parts.compact_blank.join(' ')
+      end
+    end
+  end
+end
