+++ modules/accredited_representative_portal/app/serializers/accredited_representative_portal/power_of_attorney_request_serializer/organization_power_of_attorney_holder_serializer.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module AccreditedRepresentativePortal
+  class PowerOfAttorneyRequestSerializer
+    class OrganizationPowerOfAttorneyHolderSerializer < ApplicationSerializer
+      attribute :type do
+        'veteran_service_organization'
+      end
+
+      attribute :name
+    end
+  end
+end
